# specs/openapi.gateway.yaml
...
openapi: 3.1.0
info:
  title: AI Model Serving Gateway
  description: "\nvLLM 기반 LLM·Embedding·Risk 런타임을 하나의 OpenAI 호환 API로 제공합니다.\n이 문서의\
    \ 한도·파라미터 값은 실제 배포 설정에서 자동으로 만들어집니다.\n지금 적용 중인 값은 `GET /v1/models`에서 확인하세요.\n\n\
    ## 첫 호출\n\n`$GATEWAY`는 이 문서를 열고 있는 주소입니다.\n\n```bash\ncurl -X POST \"$GATEWAY/v1/chat/completions\"\
    \ \\\n  -H \"Authorization: Bearer $API_KEY\" \\\n  -H \"Content-Type: application/json\"\
    \ \\\n  -d '{\n    \"model\": \"local-main\",\n    \"messages\": [{\"role\": \"\
    user\", \"content\": \"안녕하세요\"}]\n  }'\n```\n\n응답은 OpenAI chat completion 형식입니다.\
    \ 사용할 수 있는 모델과 조정 가능한 파라미터는\n`GET /v1/models`의 `request_parameters`가 알려줍니다.\n\n\
    ## 인증\n\n- **bearerAuth** — `/v1/*` 사용자 API: `Authorization: Bearer <API_KEY>`\n\
    - **adminBearerAuth** — `/ready`, `/metrics`, `/admin/*`: `Authorization: Bearer\
    \ <ADMIN_API_KEY>`\n"
  contact:
    name: AI Model Serving Platform 운영
  version: 0.0.1
paths:
...
  /v1/models:
    get:
      tags:
      - Models
      summary: 사용 가능한 모델 목록
      description: 'Gateway가 외부 호출자에게 노출하는 logical model id, capability, 사용자 조정 가능
        request parameter 목록입니다. catalog 성격의 엔드포인트이므로 vLLM 로딩 상태와 무관하게 항상 목록을 반환합니다.
        현재 serving 가능 여부는 `/ready`의 `status`와 `not_ready_dependencies`를 확인하세요. 클라이언트
        UI는 각 item의 `request_parameters`를 읽어 모델별 입력 form을 구성할 수 있습니다. 이미지·오디오·비디오
        입력의 개수·용량 한도는 `request_limits`에 modality별로 실립니다 -- 조정 가능한 파라미터가 아니라 콘텐츠 제약이라
        자리를 나눕니다. 두 값 모두 활성 main-model 프로필을 따르므로, 클라이언트는 복제해 두지 말고 요청 시점에 읽어야 합니다.


        조정할 수 있는 파라미터는 모델마다 다릅니다. 파라미터 이름과 허용 범위(`min`, `max`, `max_items`, `allowed`
        등)는 `GET /v1/models`의 `request_parameters`가 함께 돌려줍니다. 모델별 입력 폼을 만든다면 그 값을
        읽어서 구성하세요. 이 문서 화면에 보이는 예시 값은 화면용 preset이며 Gateway 기본값이 아닙니다.


        ### `local-main`은 이름만 고정된 창구입니다


        `local-main`은 고정된 하나의 모델이 아니라 현재 활성화된 메인 모델 프로필의 공개 이름입니다. 프로필을 바꾸면 받을 수 있는
        입력 종류, 토큰 한도, 허용 파라미터 목록이 함께 바뀌고 `/v1/models` 응답이 즉시 그 값을 반영합니다.


        | 프로필 ID | 이름 | 호환성 |

        |---|---|---|

        | `gemma4-26b-a4b-fp8` | Gemma 4 26B A4B FP8 | `verified` |

        | `gemma4-12b-unified-fp8` | Gemma 4 12B Unified FP8 | `verified` |

        | `gemma4-e4b-it` | Gemma 4 E4B IT | `verified` |

        | `qwen2.5-omni-7b-thinker` | Qwen 2.5 Omni 7B (Thinker) | `verified` |


        호환성이 `verified`가 아닌 프로필로 전환하려면 `POST /admin/main-model/switch`에 `confirm_unverified:
        true`가 필요합니다. 각 프로필의 근거는 `GET /admin/main-model/profiles`가 반환합니다.'
      operationId: listModels
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: ./schemas/model_list_response.schema.json
        '401':
          description: '인증 실패


            이 status로 올 수 있는 `error.code`:

            - `UNAUTHORIZED` — API 키 또는 인증 정보가 없거나 유효하지 않다. (재시도 불가)'
          content:
            application/json:
              schema:
                allOf:
                - $ref: ./schemas/common_error.schema.json
                - type: object
                  properties:
                    error:
                      type: object
                      properties:
                        code:
                          enum:
                          - UNAUTHORIZED
      security:
      - bearerAuth: []
      x-response-contract-schema: model_list_response.schema.json
...
  /v1/embeddings:
    post:
      tags:
      - Embeddings
      summary: Embedding vector 생성
      description: '`local-embed` 및 `local-embed-ko`를 통해 텍스트의 embedding vector를 생성합니다.
        요청 파라미터는 Gateway contract로 검증하며, 지원하지 않는 파라미터는 차단합니다.


        텍스트 embedding API입니다. 모델마다 출력 차원이 고정되어 있고, Gateway가 요청 차원과 응답 차원이 일치하는지 확인합니다.


        | 모델 | 기본 차원 | 허용 차원 | encoding_format |

        |---|---|---|---|

        | `local-embed` | `768` | `768` | `float`, `base64` |

        | `local-embed-ko` | `1024` | `1024` | `float`, `base64` |


        - `base64`는 little-endian float32 배열을 인코딩한 문자열로 반환됩니다. openai-python은 numpy가
        설치되어 있으면 이 형식을 기본으로 요청하므로, 형식을 지정하지 않은 공식 SDK 호출도 그대로 동작합니다.

        - `input`은 문자열 또는 비어 있지 않은 문자열 배열입니다. 토큰 id 배열은 받지 않습니다.'
      operationId: createEmbedding
      requestBody:
        content:
          application/json:
            schema:
              $ref: ./schemas/embedding_request.schema.json
            examples:
              basic:
                summary: 기본 요청 (단일 입력)
                value:
                  model: local-embed
                  input:
                  - 임베딩할 텍스트 예시입니다.
              truncate_prompt_tokens:
                summary: 긴 입력 truncation
                value:
                  model: local-embed
                  input:
                  - 매우 긴 문서 텍스트를 임베딩할 때 truncate_prompt_tokens로 최대 토큰 수를 제한합니다.
                  truncate_prompt_tokens: 512
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: ./schemas/embedding_response.schema.json
        '401':
          description: 'API Bearer token 필요


            이 status로 올 수 있는 `error.code`:

            - `UNAUTHORIZED` — API 키 또는 인증 정보가 없거나 유효하지 않다. (재시도 불가)'
          content:
            application/json:
              schema:
                allOf:
                - $ref: ./schemas/common_error.schema.json
                - type: object
                  properties:
                    error:
                      type: object
                      properties:
                        code:
                          enum:
                          - UNAUTHORIZED
        '422':
          description: 'Validation Error


            이 status로 올 수 있는 `error.code`:

            - `MODEL_CAPABILITY_MISMATCH` — 요청이 활성 모델이 지원하지 않는 기능을 요구한다(예 - 비활성 모달리티).
            (재시도 불가)

            - `VALIDATION_ERROR` — 요청이 Gateway 입력 계약 검증을 통과하지 못했다. (재시도 불가)'
          content:
            application/json:
              schema:
                allOf:
                - $ref: ./schemas/common_error.schema.json
                - type: object
                  properties:
                    error:
                      type: object
                      properties:
                        code:
                          enum:
                          - MODEL_CAPABILITY_MISMATCH
                          - VALIDATION_ERROR
        '413':
          description: '요청 body가 너무 큼


            이 status로 올 수 있는 `error.code`:

            - `REQUEST_TOO_LARGE` — 요청 body 가 허용 한도를 초과했다. (재시도 불가)'
          content:
            application/json:
              schema:
                allOf:
                - $ref: ./schemas/common_error.schema.json
                - type: object
                  properties:
                    error:
                      type: object
                      properties:
                        code:
                          enum:
                          - REQUEST_TOO_LARGE
        '500':
          description: '내부 서버 오류


            이 status로 올 수 있는 `error.code`:

            - `INTERNAL_ERROR` — Gateway 내부 처리 중 예기치 못한 오류가 발생했다. (재시도 불가)'
          content:
            application/json:
              schema:
                allOf:
                - $ref: ./schemas/common_error.schema.json
                - type: object
                  properties:
                    error:
                      type: object
                      properties:
                        code:
                          enum:
                          - INTERNAL_ERROR
        '429':
          description: 'Upstream rate limit


            이 status로 올 수 있는 `error.code`:

            - `RATE_LIMITED` — upstream rate limit에 도달했다. (재시도 가능)'
          content:
            application/json:
              schema:
                allOf:
                - $ref: ./schemas/common_error.schema.json
                - type: object
                  properties:
                    error:
                      type: object
                      properties:
                        code:
                          enum:
                          - RATE_LIMITED
        '503':
          description: 'Runtime 또는 dependency를 사용할 수 없음


            이 status로 올 수 있는 `error.code`:

            - `CIRCUIT_OPEN` — 연속 실패로 upstream 서킷이 열려 일시적으로 차단 중이다. (재시도 가능)

            - `MODEL_UNAVAILABLE` — 대상 모델 런타임을 현재 사용할 수 없다(미기동·축출 등). (재시도 가능)

            - `QUEUE_TIMEOUT` — 동시 처리 한도로 대기열에서 시간이 초과됐다. (재시도 가능)'
          content:
            application/json:
              schema:
                allOf:
                - $ref: ./schemas/common_error.schema.json
                - type: object
                  properties:
                    error:
                      type: object
                      properties:
                        code:
                          enum:
                          - CIRCUIT_OPEN
                          - MODEL_UNAVAILABLE
                          - QUEUE_TIMEOUT
        '504':
          description: 'Upstream timeout


            이 status로 올 수 있는 `error.code`:

            - `UPSTREAM_TIMEOUT` — upstream 런타임 응답이 제한 시간을 초과했다. (재시도 가능)'
          content:
            application/json:
              schema:
                allOf:
                - $ref: ./schemas/common_error.schema.json
                - type: object
                  properties:
                    error:
                      type: object
                      properties:
                        code:
                          enum:
                          - UPSTREAM_TIMEOUT
        '502':
          description: 'Upstream 오류 또는 유효하지 않은 upstream 응답


            이 status로 올 수 있는 `error.code`:

            - `UPSTREAM_ERROR` — upstream 런타임이 오류를 반환했거나 통신에 실패했다. (재시도 가능)

            - `UPSTREAM_RESPONSE_INVALID` — upstream 응답의 envelope, 필드 또는 detector
            출력이 계약을 만족하지 못했다. (재시도 가능)'
          content:
            application/json:
              schema:
                allOf:
                - $ref: ./schemas/common_error.schema.json
                - type: object
                  properties:
                    error:
                      type: object
                      properties:
                        code:
                          enum:
                          - UPSTREAM_ERROR
                          - UPSTREAM_RESPONSE_INVALID
      security:
      - bearerAuth: []
      x-contract-schema: embedding_request.schema.json
      x-response-contract-schema: embedding_response.schema.json
...
components:
...
  securitySchemes:
    adminBearerAuth:
      type: http
      description: /ready, /metrics 같은 운영자 endpoint에 사용하는 admin Bearer token입니다.
      scheme: bearer
    bearerAuth:
      type: http
      description: Gateway API 또는 내부 Risk Adapter 호출에 사용하는 Bearer token입니다.
      scheme: bearer
...